Changed When Assets are Loaded to Speed Up Intro
Previously, all assets would be loaded in the preload() function. Additionally, the model would immediately be loaded during setup. These two meant that the program would not only take a long time to even start but also it would freeze for a couple seconds in the beginning before anything even happens.

Now, nothing is loaded when the program starts aside from the main logo (which is pretty quick loading). Here is how everything else is loaded:
- The model will only be loaded once the user presses ENTER to go to the "Adjust Camera" scene.
- Songs info will only be loaded when entering the "Select Level" scene.
- Sound effects and game images will only be loaded when entering the "Game Scene" or "Tutorial" scenes.